split out neon pipeline helper methods into separate module
